requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
ccimport==0.4.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.2.1
cumm-cu124==0.7.11
docker-pycreds==0.4.0
filelock==3.13.1
fire==0.7.0
fpsample==0.3.3
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
gitpython==3.1.44
hydra-core==1.3.2
idna==3.4
jinja2==3.1.4
joblib==1.5.1
lark==1.2.2
lightning==2.5.1.post0
lightning-utilities==0.11.8
loguru==0.7.3
markupsafe==2.1.5
mpmath==1.3.0
multidict==6.4.4
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.1
pccm==0.4.16
pillow==11.0.0
platformdirs==4.3.8
portalocker==2.10.1
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pybind11==2.13.6
pydantic==2.11.5
pydantic-core==2.33.2
pytorch-lightning==2.5.1.post0
pyyaml==6.0.2
requests==2.28.1
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
spconv-cu124==2.3.8
sympy==1.13.1
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.5.0+cu124
torchmetrics==1.0.3
torchvision==0.20.0+cu124
tqdm==4.66.5
triton==3.1.0
typing-extensions==4.12.2
typing-inspection==0.4.1
urllib3==1.26.13
volumentations==0.1.8
wandb==0.19.11
yarl==1.20.0
